headers added, clean up, input_ex moved
