+++ nemo/collections/nlp/data/datasets/sgd_dataset/evaluate.py
@@ -1,5 +1,6 @@
-# coding=utf-8
-# Copyright 2020 The Google Research Authors.
+# =============================================================================
+# Copyright 2020 NVIDIA. All Rights Reserved.
+# Copyright 2019 The Google Research Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,6 +13,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+# =============================================================================
 
 """Evaluate predictions JSON file, w.r.t. ground truth file."""
